feat: move registry related helpers to registry.py
